requirements.txt
Flask~=2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
python-dotenv==1.0.0
psycopg2-binary==2.9.11 # for Postgres

Werkzeug~=3.1.3
SQLAlchemy~=2.0.44
alembic~=1.17.0